refactor(transpile): split go backend orchestration and formatting
