+++ compiler/transpile/ir.go
@@ -1,0 +1,24 @@
+package transpile
+
+type goFileIR struct {
+	PackageName string
+	Imports     []goImportIR
+}
+
+type goImportIR struct {
+	Alias string
+	Path  string
+}
+
+func lowerGoFileIR(packageName string, imports map[string]string) goFileIR {
+	file := goFileIR{PackageName: packageName}
+	paths := sortedImportPaths(imports)
+	file.Imports = make([]goImportIR, 0, len(paths))
+	for _, importPath := range paths {
+		file.Imports = append(file.Imports, goImportIR{
+			Alias: imports[importPath],
+			Path:  importPath,
+		})
+	}
+	return file
+}
